MP: cleanup references
