+++ ExtLibs/ZedGraph/ZedGraph.csproj
@@ -120,9 +120,6 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="..\Utilities\MeasureString.cs">
-      <Link>MeasureString.cs</Link>
-    </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ReversibleFrame.cs" />
     <Compile Include="ZedGraph\ArrowObj.cs" />
