+++ MissionPlanner.csproj
@@ -1049,18 +1049,22 @@
     <Compile Include="Utilities\GMapMarkerRallyPt.cs" />
     <Compile Include="Utilities\ICorrections.cs" />
     <Compile Include="Utilities\ImageProjection.cs" />
+    <Compile Include="Utilities\LangUtility.cs" />
     <Compile Include="Utilities\LineOffset.cs" />
     <Compile Include="Utilities\ListExtension.cs" />
     <Compile Include="Utilities\LogAnalyzer.cs" />
     <Compile Include="NoFly\NoFly.cs" />
+    <Compile Include="Utilities\MeasureString.cs" />
     <Compile Include="Utilities\NGEN.cs" />
     <Compile Include="Utilities\OpticalFlow.cs" />
     <Compile Include="Utilities\ParameterMetaDataRepositoryAPM.cs" />
     <Compile Include="Utilities\ParameterMetaDataRepositoryPX4.cs" />
     <Compile Include="Utilities\Firmware.cs" />
     <Compile Include="Utilities\Proximity.cs" />
+    <Compile Include="Utilities\Speech.cs" />
     <Compile Include="Utilities\Tools.cs" />
     <Compile Include="Utilities\nmea.cs" />
+    <Compile Include="Utilities\Tracking.cs" />
     <Compile Include="Utilities\ubx_m8p.cs" />
     <Compile Include="Utilities\UDPMavlinkShim.cs" />
     <Compile Include="Utilities\UDPVideoShim.cs" />
