+++ Utilities/MeasureString.cs
